Show the actual signs of modifiers even when ACO disabled
I.e. don't show 25% city attack with a positive sign when it actually gets subtracted from the defender. Otherwise, we get e.g.
+25% city attack
-25% vs. Melee
both in green if both benefit the attacker. Looks like they would cancel out. Put differently, the BtS coloring scheme doesn't work well in a mod where negative modifiers are commonplace.

To avoid confusion with generic modifiers of the attacker (which apply to the attacker's combat strength), I'm showing the one-line unit help of the defender as a separator between modifiers that apply to the attacker and modifiers that apply to the defender (like in ACO).